Funciona be el checkkout amb una tenda unicament falta provar amb mes de una a la vegada

+++ src/BUSSINESS/ENTITIES/Producte.java
@@ -136,7 +136,7 @@ public class Producte {
      *
      * @return Valor de l'IVA del producte amb impost reduït.
      */
-    public int getIva() {
+    public float getIva() {
         return 0;
     }
